Do smart things with the permissions/roles.

+++ src/AppBundle/Controller/ArtworkCategoryController.php
@@ -48,6 +48,10 @@ class ArtworkCategoryController extends Controller
      */
     public function newAction(Request $request)
     {
+        if( ! $this->isGranted('ROLE_CONTENT_ADMIN')) {
+            $this->addFlash('danger', 'You must login to access this page.');
+            return $this->redirect($this->generateUrl('fos_user_security_login'));
+        }
         $artworkCategory = new ArtworkCategory();
         $form = $this->createForm(ArtworkCategoryType::class, $artworkCategory);
         $form->handleRequest($request);
@@ -94,6 +98,10 @@ class ArtworkCategoryController extends Controller
      */
     public function editAction(Request $request, ArtworkCategory $artworkCategory)
     {
+        if( ! $this->isGranted('ROLE_CONTENT_ADMIN')) {
+            $this->addFlash('danger', 'You must login to access this page.');
+            return $this->redirect($this->generateUrl('fos_user_security_login'));
+        }
         $editForm = $this->createForm(ArtworkCategoryType::class, $artworkCategory);
         $editForm->handleRequest($request);
 
@@ -120,6 +128,10 @@ class ArtworkCategoryController extends Controller
      */
     public function deleteAction(Request $request, ArtworkCategory $artworkCategory)
     {
+        if( ! $this->isGranted('ROLE_CONTENT_ADMIN')) {
+            $this->addFlash('danger', 'You must login to access this page.');
+            return $this->redirect($this->generateUrl('fos_user_security_login'));
+        }
         $em = $this->getDoctrine()->getManager();
         $em->remove($artworkCategory);
         $em->flush();
