+++ src/AppBundle/Controller/PersonController.php
@@ -101,6 +101,10 @@ class PersonController extends Controller {
      * @param Request $request
      */
     public function newAction(Request $request) {
+        if( ! $this->isGranted('ROLE_CONTENT_ADMIN')) {
+            $this->addFlash('danger', 'You must login to access this page.');
+            return $this->redirect($this->generateUrl('fos_user_security_login'));
+        }
         $person = new Person();
         $form = $this->createForm(PersonType::class, $person);
         $form->handleRequest($request);
@@ -145,6 +149,10 @@ class PersonController extends Controller {
      * @param Person $person
      */
     public function editAction(Request $request, Person $person) {
+        if( ! $this->isGranted('ROLE_CONTENT_ADMIN')) {
+            $this->addFlash('danger', 'You must login to access this page.');
+            return $this->redirect($this->generateUrl('fos_user_security_login'));
+        }
         $editForm = $this->createForm(PersonType::class, $person);
         $editForm->handleRequest($request);
 
@@ -170,6 +178,10 @@ class PersonController extends Controller {
      * @param Person $person
      */
     public function deleteAction(Request $request, Person $person) {
+        if( ! $this->isGranted('ROLE_CONTENT_ADMIN')) {
+            $this->addFlash('danger', 'You must login to access this page.');
+            return $this->redirect($this->generateUrl('fos_user_security_login'));
+        }
         $em = $this->getDoctrine()->getManager();
         $em->remove($person);
         $em->flush();
@@ -187,6 +199,10 @@ class PersonController extends Controller {
      * @param Person $person
      */
     public function addMediaAction(Request $request, Person $person) {
+        if( ! $this->isGranted('ROLE_CONTENT_ADMIN')) {
+            $this->addFlash('danger', 'You must login to access this page.');
+            return $this->redirect($this->generateUrl('fos_user_security_login'));
+        }
         $em = $this->getDoctrine()->getManager();
         $repo = $em->getRepository('AppBundle:MediaFile');
         $q = $request->query->get('q');
@@ -230,6 +246,10 @@ class PersonController extends Controller {
      * @param Person $person
      */
     public function removeMediaAction(Request $request, Person $person) {
+        if( ! $this->isGranted('ROLE_CONTENT_ADMIN')) {
+            $this->addFlash('danger', 'You must login to access this page.');
+            return $this->redirect($this->generateUrl('fos_user_security_login'));
+        }
         $paginator = $this->get('knp_paginator');
         $results = $paginator->paginate($person->getMediaFiles(), $request->query->getInt('page', 1), 25);
 
@@ -265,6 +285,10 @@ class PersonController extends Controller {
      * @param Person $person
      */
     public function projectContributionsAction(Request $request, Person $person) {
+        if( ! $this->isGranted('ROLE_CONTENT_ADMIN')) {
+            $this->addFlash('danger', 'You must login to access this page.');
+            return $this->redirect($this->generateUrl('fos_user_security_login'));
+        }
         $form = $this->createForm(ProjectContributionsType::class, $person, array(
             'person' => $person,
         ));
@@ -292,6 +316,10 @@ class PersonController extends Controller {
      * @param Person $person
      */
     public function artworkContributionsAction(Request $request, Person $person) {
+        if( ! $this->isGranted('ROLE_CONTENT_ADMIN')) {
+            $this->addFlash('danger', 'You must login to access this page.');
+            return $this->redirect($this->generateUrl('fos_user_security_login'));
+        }
         $form = $this->createForm(ArtworkContributionsType::class, $person, array(
             'person' => $person,
         ));
